+++ sdk/src/rhesis/sdk/metrics/providers/native/factory.py
@@ -1,10 +1,10 @@
 from typing import List
 
 from rhesis.sdk.metrics.base import BaseMetric, BaseMetricFactory
-from rhesis.sdk.metrics.providers.native.prompt_categorical_metric import (
+from rhesis.sdk.metrics.providers.native.prompt_metric_categorical import (
     RhesisPromptMetricCategorical,
 )
-from rhesis.sdk.metrics.providers.native.prompt_numeric_metric import RhesisPromptMetricNumeric
+from rhesis.sdk.metrics.providers.native.prompt_metric_numeric import RhesisPromptMetricNumeric
 
 
 class RhesisMetricFactory(BaseMetricFactory):
